refactor(#41): rename dependency range policy files to kebab-case
Addresses review feedback that file names should not use camelCase.
Renames scripts/ci/dependencyRangePolicy.ts and its unit test to
kebab-case and updates the imports, JSDoc link, and CONTRIBUTING
reference accordingly.
